Removed compilation warning
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7502 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_draw.cxx
@@ -234,7 +234,7 @@ void fl_draw(
     for (p = str, lines=0; p;) {
       e = fl_expand_text(p, buf, MAXBUF, w - symtotal, buflen, width, 
                          align&FL_ALIGN_WRAP, draw_symbols);
-      if (strw<width) strw = width;
+      if (strw<width) strw = (int)width;
       lines++;
       if (!*e || (*e == '@' && e[1] != '@' && draw_symbols)) break;
       p = e;
